+++ app-backend/datamodel/src/main/scala/AnnotationLabel.scala
@@ -17,7 +17,8 @@ final case class AnnotationLabel(
     createdBy: String,
     geometry: Option[Projected[Geometry]],
     annotationProjectId: UUID,
-    annotationTaskId: UUID
+    annotationTaskId: UUID,
+    description: Option[String] = None
 )
 
 @JsonCodec
@@ -25,12 +26,14 @@ final case class AnnotationLabelProperties(
     createdAt: Timestamp,
     createdBy: String,
     annotationProjectId: UUID,
-    annotationTaskId: UUID
+    annotationTaskId: UUID,
+    description: Option[String] = None
 )
 
 final case class AnnotationLabelPropertiesCreate(
     annotationProjectId: UUID,
-    annotationTaskId: UUID
+    annotationTaskId: UUID,
+    description: Option[String] = None
 )
 
 @JsonCodec
@@ -46,6 +49,7 @@ final case class AnnotationLabelWithClasses(
     geometry: Option[Projected[Geometry]],
     annotationProjectId: UUID,
     annotationTaskId: UUID,
+    description: Option[String] = None,
     annotationLabelClasses: List[UUID]
 ) extends GeoJSONSerializable[AnnotationLabelWithClasses.GeoJSON] {
   def toGeoJSONFeature = AnnotationLabelWithClasses.GeoJSON(
@@ -56,7 +60,8 @@ final case class AnnotationLabelWithClasses(
       this.createdBy,
       this.annotationProjectId,
       this.annotationTaskId,
-      this.annotationLabelClasses
+      this.annotationLabelClasses,
+      this.description
     )
   )
 
@@ -97,7 +102,8 @@ object AnnotationLabelWithClasses {
 
   final case class Create(
       geometry: Option[Projected[Geometry]],
-      annotationLabelClasses: List[UUID]
+      annotationLabelClasses: List[UUID],
+      description: Option[String] = None
   ) {
     def toAnnotationLabelWithClasses(
         annotationProjectId: UUID,
@@ -112,6 +118,7 @@ object AnnotationLabelWithClasses {
         geometry,
         annotationProjectId,
         annotationTaskId,
+        description,
         annotationLabelClasses
       )
     }
@@ -133,7 +140,8 @@ object AnnotationLabelWithClasses {
       : AnnotationLabelWithClasses.Create = {
       AnnotationLabelWithClasses.Create(
         geometry,
-        properties.annotationLabelClasses
+        properties.annotationLabelClasses,
+        properties.description
       )
     }
   }
@@ -203,7 +211,8 @@ object AnnotationLabelWithClasses {
 
 @JsonCodec
 final case class AnnotationLabelWithClassesPropertiesCreate(
-    annotationLabelClasses: List[UUID]
+    annotationLabelClasses: List[UUID],
+    description: Option[String] = None
 )
 
 @JsonCodec
@@ -212,7 +221,8 @@ final case class AnnotationLabelWithClassesProperties(
     createdBy: String,
     annotationProjectId: UUID,
     annotationTaskId: UUID,
-    annotationLabelClasses: List[UUID]
+    annotationLabelClasses: List[UUID],
+    description: Option[String] = None
 )
 
 final case class StacGeoJSONFeatureCollection(